dont comment out the json, use packer env variables to determine if vmtools should be installed

+++ packer/scripts/common/vmtools.sh
@@ -1,23 +1,27 @@
 #!/bin/bash
 
-if [ $PACKER_BUILDER_TYPE == 'virtualbox' ]; then
-    mkdir /tmp/vbox
-    VER=$(cat /home/vagrant/.vbox_version)
-    mount -o loop /home/vagrant/VBoxGuestAdditions_$VER.iso /tmp/vbox 
-    sh /tmp/vbox/VBoxLinuxAdditions.run
-    umount /tmp/vbox
-    rmdir /tmp/vbox
-    rm /home/vagrant/*.iso
-fi
+if [ $VM_TOOLS != 'disable' ]; then
+    if [ $PACKER_BUILDER_TYPE == 'virtualbox' ]; then
+        mkdir /tmp/vbox
+        VER=$(cat /home/vagrant/.vbox_version)
+        mount -o loop /home/vagrant/VBoxGuestAdditions_$VER.iso /tmp/vbox
+        sh /tmp/vbox/VBoxLinuxAdditions.run
+        umount /tmp/vbox
+        rmdir /tmp/vbox
+        rm /home/vagrant/*.iso
+    fi
 
-if [ $PACKER_BUILDER_TYPE == 'vmware' ]; then
-    mkdir /tmp/vmfusion
-    mkdir /tmp/vmfusion-archive
-    mount -o loop /home/vagrant/linux.iso /tmp/vmfusion
-    tar xzf /tmp/vmfusion/VMwareTools-*.tar.gz -C /tmp/vmfusion-archive
-    /tmp/vmfusion-archive/vmware-tools-distrib/vmware-install.pl --default
-    umount /tmp/vmfusion
-    rm -rf  /tmp/vmfusion
-    rm -rf  /tmp/vmfusion-archive
-    rm /home/vagrant/*.iso
+    if [ $PACKER_BUILDER_TYPE == 'vmware' ]; then
+        mkdir /tmp/vmfusion
+        mkdir /tmp/vmfusion-archive
+        mount -o loop /home/vagrant/linux.iso /tmp/vmfusion
+        tar xzf /tmp/vmfusion/VMwareTools-*.tar.gz -C /tmp/vmfusion-archive
+        /tmp/vmfusion-archive/vmware-tools-distrib/vmware-install.pl --default
+        umount /tmp/vmfusion
+        rm -rf  /tmp/vmfusion
+        rm -rf  /tmp/vmfusion-archive
+        rm /home/vagrant/*.iso
+    fi
+else
+    echo "Building the box without Guest Additions..."
 fi
